messages: avoid converting ceph errors on Windows
libclient and libcephfs return CEPHFS_E* errors, which are basically
Linux errno codes. If we convert mds errors to host errno values, we
end up mixing error codes.

For Windows, we'll preserve the original error value, which is expected
to be a Linux (CEPHFS_E*) error.

It may be worth doing the same for other platforms eventually.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>